+++ backend/src/test/java/com/pickpick/channel/ui/ChannelControllerTest.java
@@ -1,4 +1,4 @@
-package com.pickpick.controller;
+package com.pickpick.channel.ui;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
@@ -9,6 +9,7 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.response
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.pickpick.auth.support.JwtTokenProvider;
+import com.pickpick.config.RestDocsTestSupport;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
