+++ backend/src/test/java/com/pickpick/channel/ui/ChannelSubscriptionControllerTest.java
@@ -1,4 +1,4 @@
-package com.pickpick.controller;
+package com.pickpick.channel.ui;
 
 
 import static org.mockito.ArgumentMatchers.any;
@@ -15,6 +15,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.pickpick.auth.support.JwtTokenProvider;
 import com.pickpick.channel.ui.dto.ChannelOrderRequest;
 import com.pickpick.channel.ui.dto.ChannelSubscriptionRequest;
+import com.pickpick.config.RestDocsTestSupport;
 import java.util.List;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.BeforeEach;
