+++ backend/src/test/java/com/pickpick/config/RestDocsConfiguration.java
@@ -1,4 +1,4 @@
-package com.pickpick.controller;
+package com.pickpick.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -7,7 +7,7 @@ import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
 import org.springframework.restdocs.operation.preprocess.Preprocessors;
 
 @TestConfiguration
-class RestDocsConfiguration {
+public class RestDocsConfiguration {
 
     @Bean
     public RestDocumentationResultHandler write() {
